Verification code

+++ mesh_spot.cpp
@@ -94,11 +94,9 @@ namespace project{
         std::cout << std::endl;
     };
 
-	std::vector<std::vector<double>> transform_matrix(const std::vector<double>& vector_init, const double& nb_rows){
+	std::vector<std::vector<double>> transform_matrix(const std::vector<double>& vector_init, const size_t& nb_rows){
 	
 	double endv; 
-	//std::vector<double> upper_bound(vector_init);
-	//std::vector<double> lower_bound(vector_init);
